predicts/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
auto-py-to-exe==2.48.1
awscli==1.44.16
boto3==1.42.26
botocore==1.42.26
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docker==7.1.0
docutils==0.19
Eel==0.18.2
fastapi==0.128.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
Jinja2==3.1.6
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.11.0.86
packaging==25.0
pefile==2024.8.26
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
ruamel.yaml==0.19.1
s3transfer==0.16.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.252
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
voluptuous==0.16.0
yolo==0.3.1
zope.event==6.1
zope.interface==8.2
python-dotenv==1.0.1
